LIBITD-1648. Added etc/config.yaml file for application config
Added an "etc/config.yaml" file for storing application configuration
information that is not server-specific, but that may change in the
future.

Added the "pyyaml" library to the dependencies to enable YAML parsing.
YAML was chosen for the configuration file, as it allows explanatory
comments (something JSON does not).

Removed "CIRCREQUESTS_SOURCE_KEY_FIELD" from the "env_example" file,
moving the field into the "etc/config.yaml" file as it is not
server-specific

Modified the JobConfig base class to store the application configuration
in an "application_configuration" entry.

Replaced references to "CIRCREQUESTS_SOURCE_KEY_FIELD" with references
to the application configuration in the job config.

Moved the "storage/etc" directory to "etc/", as the "storage" directory
no longer seemed to be a good place to store the application
configuration.

Updated the tests and associated test resource files to use "barcode",
instead of "item" as the "source_key_field". This was done to make the
examples more like the actual responses from Aleph.

Removed the "source_key_field" check from the
"validate_job_preconditions" step, as the value is now in the
"etc/config.yaml" file (as opposed to the ".env" file), and so is
not likely to be missing or invalid.

https://issues.umd.edu/browse/LIBITD-1648