LIBITD-1671. Made "invoke" tasks more granular
Added "invoke" tasks for cleaning/resetting "items" as well as
"circrequests", and added additional tasks to make them more
granular.

The "clean" task no longer calls "reset", and added "reset-and-clean"
task so that these tasks would be called in the right order (if "clean"
is called before "reset", the file pointed to be the "last success" file
would not be deleted).

https://issues.umd.edu/browse/LIBITD-1671