LIBITD-1676. Modified end_time to be derived from source response
Instead of generating the the end_time using the current time, and
sending to the source, modified the steps so that:

1) For the first iteration, the end_time is not provided in the
request.

2) The end_time is retrieved from the source response, and provided in
the next iteration (if there is one).

The enables the source to (a) provide the end_time, and (b) removes
problems such as out-of-sync clocks on the middleware server and the
source, as well as any timezone issues.

https://issues.umd.edu/browse/LIBITD-1676