LIBITD-1681. Enabled null values to be replaced with sentinel value
For "items" updates, Aleph will always send information about all the
fields for a record, using "null" for fields where there is no
information.

CaiaSoft does not automatically "clear" empty fields. So in order to
ensure that the data from Aleph and the data in CaiaSoft is in sync,
"null" values in "items" updates should be modified by the middleware
to the sentinel value "CLEARFIELD*". This will cause CaiaSoft to always
clear the field, regardless of its current value.

In the "parse_items" method of "caia/items/steps/create_dest_request.py",
replaced the unused "suppress_null_values" parameter with
"convert_null_values", and, when True, modified the method to replace
null values with the CaiaSoft sentinel value.

Updated the test cases.

https://issues.umd.edu/browse/LIBITD-1681