LIBITD-1735. Changed code to immediately store destination response
Changed the code to immediately store the destination response, instead
of first parsing it as JSON and then storing.

This change was motivated by the destination not always returning
parseable JSON, and so we want to record the response before parsing
so we know what the response was for troubleshooting purposes.

https://issues.umd.edu/browse/LIBITD-1735