LIBITD-1885. Pin "mbtest" version to v2.5.1 to fix Jenkins builds
Pinned the "mbtest" version to v2.5.1, due a breaking change in v2.5.2
of the library that causes the tests to fail.

Specifically, any test that calls the "get_actual_requests" method will
fail with the following error:

```
TypeError: 'generator' object is not subscriptable
```

This is causes by a change to the return value of the method in the
library.

Added an ADR describing the context of this decision.

https://issues.umd.edu/browse/LIBITD-1885